# docs/automation/openapi.yaml
...
openapi: 3.0.1
info:
  title: Gravitee.io AM - Automation API
  description: "Declarative, GitOps-style management of Access Management resources."
  version: 4.12.0-alpha.5
servers:
- url: /management/automation
security:
- BearerAuth: []
- BasicAuth: []
...
paths:
...
  /organizations/{orgId}/environments/{envId}/domains/{domainKey}/certificates:
    get:
      tags:
      - Certificates
      - Domains
      summary: List a domain's certificates
      description: Returns all certificates managed by the Automation API under the
        domain.
      operationId: automationListCertificates
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: envId
        in: path
        required: true
        schema:
          type: string
      - name: domainKey
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: List of certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutomationCertificate"
        "403":
          description: Permission denied
        default:
          description: Unexpected error
    put:
      tags:
      - Certificates
      - Domains
      summary: Create or update a certificate
      description: Idempotent create-or-update. Uses the key field in the body to
        identify the certificate within the domain.
      operationId: automationCreateOrUpdateCertificate
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: envId
        in: path
        required: true
        schema:
          type: string
      - name: domainKey
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationCertificate"
        required: true
      responses:
        "200":
          description: The created or updated certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationCertificate"
        "403":
          description: Permission denied
        default:
          description: Unexpected error
  /organizations/{orgId}/environments/{envId}/domains/{domainKey}/certificates/{certKey}:
    get:
      tags:
      - Certificates
      - Domains
      summary: Get a certificate
      description: Retrieves a single certificate by its key.
      operationId: automationGetCertificate
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: envId
        in: path
        required: true
        schema:
          type: string
      - name: domainKey
        in: path
        required: true
        schema:
          type: string
      - name: certKey
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationCertificate"
        "403":
          description: Permission denied
        default:
          description: Unexpected error
    delete:
      tags:
      - Certificates
      - Domains
      summary: Delete a certificate
      operationId: automationDeleteCertificate
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
      - name: envId
        in: path
        required: true
        schema:
          type: string
      - name: domainKey
        in: path
        required: true
        schema:
          type: string
      - name: certKey
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Certificate deleted
        "403":
          description: Permission denied
        default:
          description: Unexpected error
...
components:
  schemas:
...
    AutomationCertificate:
      required:
      - key
      type: object
      properties:
        configuration:
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        expiresAt:
          type: string
          format: date-time
          readOnly: true
        key:
          maxLength: 255
          minLength: 1
          pattern: "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
          type: string
        name:
          maxLength: 255
          minLength: 1
          type: string
        system:
          type: boolean
          description: "whether this is the domain's system certificate (immutable\
            \ after creation; when true, only key is required and the certificate\
            \ is built from domains.certificates.default.* system settings)"
        type:
          maxLength: 2147483647
          minLength: 1
          type: string
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      scheme: basic